requirements.txt
```text
torch
transformers
psutil
llama-cpp-python
```